+++ deployables/app/app/routes/auth/sign-up.tsx
@@ -2,7 +2,7 @@ import { Page } from '@/components'
 import { createTenant, dbClient } from '@/db'
 import { useIsPending } from '@/hooks'
 import { routeTitle } from '@/utils'
-import { createOrganization } from '@/workOS'
+import { createOrganization } from '@/workOS/server'
 import * as Sentry from '@sentry/react-router'
 import { getString } from '@zodiac/form-data'
 import { EmailInput, Error, Form, PrimaryButton, TextInput } from '@zodiac/ui'
