+++ deployables/app/workOS/client/index.ts
@@ -1,0 +1,1 @@
+export { Widgets } from './Widgets'
